Refactor management functions into modules
ref: #239
